Split Feature module to prevent import cycles
